requirements/testing.txt
-r base.txt

coverage==4.0.3
django-nose==1.4.3
nose==1.3.7
nose-exclude==0.4.1
pep8==1.7.0
pylint==1.5.5
mock==1.3.0
